+++ news/views/__init__.py
@@ -4,8 +4,8 @@ import logging
 from .facebook_views import FacebookPostView
 from .feeds_views import ajax_articles, articles_list, new_feed, parse_manual, parse_feed
 from .nltk_views import ajax_nltk, get_by_word_and_date, nltk_list, nltk_all, TrackedWordView
-from .twitter_views import tweets_list, stop_stream, new_tweet_stream
-from .misc_views import settings
+from .twitter_views import tweets_list
+from .settings_views import settings, stop_stream, new_tweet_stream
 
 __author__ = 'ilov3'
 logger = logging.getLogger(__name__)
